Add ignore option
LiveReload is not notfified about changes to files in the ignore list.

Usage:

    new LiveReloadPlugin({ignore: ['css_modules.js']})

Motivation:

I'm using CSS Modules, and every time I make a change to a CSS file, the .js file that loads the modules is touched. This causes LiveReload to do a full browser reload for every CSS change. With the ignore setting, LiveReload is able to hot-reload just the CSS.